# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaundryPOS API",
        "description": "Dokumentasi API LaundryPOS Ecosystem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Detail layanan",
                "operationId": "dce5b306a141bdec91eff8a15534d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail layanan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Layanan tidak ditemukan"
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Daftar Transaksi",
                "operationId": "85b7f9074f7849c472cce811af893538",
                "responses": {
                    "200": {
                        "description": "Daftar transaksi berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status-laundry": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Ambil Status Cucian Milik Pelanggan Yang Login",
                "operationId": "02a7be40bb6c61c4d749422f3125f7e1",
                "responses": {
                    "200": {
                        "description": "Data Transaksi Berhasil Diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "required": [
                    "id",
                    "service_name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "service_name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "required": [
                    "id",
                    "invoice_code",
                    "customer_id",
                    "service_id",
                    "total_price"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "invoice_code": {
                        "type": "string"
                    },
                    "customer_id": {
                        "type": "integer"
                    },
                    "service_id": {
                        "type": "integer"
                    },
                    "total_price": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string"
                    },
                    "payment_method": {
                        "type": "string"
                    },
                    "payment_status": {
                        "type": "string"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}